pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-brain"
version = "0.1.0"
description = "AI-native organizational memory system with Claude and Voyage AI"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Shu", email = "shu@anthropic.com"}
]

dependencies = [
    # FastAPI & HTTP
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "uvicorn[standard]>=0.24.0",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Pydantic
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # LLM & Embeddings
    "anthropic>=0.21.0",
    "voyageai>=0.2.0",
    "pgvector>=0.3.0",

    # Utilities
    "tenacity>=8.2.0",
    "tiktoken>=0.5.0",

    # CLI
    "typer>=0.9.0",

    # Logging
    "structlog>=23.2.0",
    "python-json-logger>=2.0.0",

    # Discord integration
    "discord.py>=2.3.0",

    # Rate limiting
    "slowapi>=0.1.9",

    # MCP server (AI-agnostic tool access)
    "fastmcp>=3.0.0",
]

[project.optional-dependencies]
calendar = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
ob = "cli.ob:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = "voyageai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "discord.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cli.*"
disallow_untyped_defs = false
warn_return_any = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
asyncio_default_test_loop_scope = "module"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
